pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kernel_writer"
version = "1.0.0"
description = "OpenEnv-compatible CUDA kernel optimization environment."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "openai>=1.0.0",
  "openenv-core>=0.2.0",
  "python-dotenv>=1.0.0",
  "gradio>=4.44.0",
  "datasets>=2.20.0",
  "trl>=0.12.0"
]

[project.scripts]
server = "server.app:main"